+++ test.c
@@ -5,50 +5,10 @@
 #include "board.h"
 #include "board.c"
 #include "moveGenerator.c"
+#include "king.c"
 
 void testKingMoves(){
-    setBoards();
-    INT board = 0;
-    INT wboard = 0;
-    INT bboard = 0;
-    piece k;
-    piece pawn;
-    k.position = 0x8000000000000000;
-    k.type = KING;
-    k.color = WHITE;
-
-    for(int i = 0; i < 64; i++){
-        if(i) k.position = k.position >> 1;
-        board = k.position;
-        wboard = k.position;
-        int moveCount = 0;
-        INT *moves = generateMoves(&k, board, wboard, bboard, &moveCount);
-        INT m = 0;
-        for(int i = 0; i < moveCount; i++){
-            m = m | moves[i];
-        }
-        free(moves);
-        printf("%lx\n", m);
-    }
-
-    // board = board | k.position;
-    // wboard = k.position;
-    // bboard = 0;
-
-    // int moveCount = 0;
-    // INT *moves = generateMoves(&k, board, wboard, bboard, &moveCount);
-
-    // INT m = 0;
-
-    // for(int i = 0; i < moveCount; i++){
-    //     printf("%llx\n", moves[i]);
-    //     m = m | moves[i];
-    // }
-
-    //Linux version
-    //printf("%lx\n", m);
-    //Mac version
-    //printf("%llx\n", m);
+    moveTest();
 }
 
 //full basic test with no other pieces
@@ -89,7 +49,9 @@ void testPawnMoves(){
     INT *moves = generateMoves(&pawn, board, wboard, bboard, &moveCount);
     printf("%d\n", moveCount);
     for(int i = 0; i < moveCount; i++){
-        printf("%lx\n", moves[i]);
+        char str[73];
+        numToPosition(moves[i], str);
+        printf("%s\n", str);
     }
     free(moves);
 }
@@ -289,30 +251,3 @@ void testQueenMoves(){
     printf("%lx\n", m);
 }
 
-//function for transforming a position into a human readable chess board format
-void numToPosition(INT pos, char *str){
-    //64 bits, 8 new lines, and one string terminator = 73 chars
-    char result[73];
-    //iterator for the string
-    int strI = 0;
-    //create a mask to shift right each iteration
-    INT mask = 0x8000000000000000;
-    char c;
-    for(int i = 0; i < 64; i++){
-        //get the bit and add it to the result
-        if(pos & (mask >> i)){
-            c = '1';
-        }else{
-            c = '0';
-        }
-        str[strI] = c;
-        strI++;
-        //if the rank should end, add a newline
-        if((i + 1) % 8 == 0){
-            str[strI] = '\n';
-            strI++;
-        }
-    }
-    //add a string terminator
-    str[strI] = '\0';
-}
