+++ testResults/fullPawnBlockTest.txt
@@ -1,0 +1,96 @@
+1
+8000000000000000
+1
+4000000000000000
+1
+2000000000000000
+1
+1000000000000000
+1
+800000000000000
+1
+400000000000000
+2
+300000000000000
+0
+0
+2
+180000000000000
+1
+40000000000000
+1
+20000000000000
+1
+10000000000000
+1
+8000000000000
+1
+4000000000000
+1
+2000000000000
+1
+1000000000000
+1
+800000000000
+0
+0
+1
+200000000000
+1
+100000000000
+1
+80000000000
+1
+40000000000
+1
+20000000000
+1
+10000000000
+1
+8000000000
+1
+4000000000
+1
+2000000000
+1
+1000000000
+1
+800000000
+1
+400000000
+1
+200000000
+1
+100000000
+1
+80000000
+1
+40000000
+1
+20000000
+1
+10000000
+1
+8000000
+1
+4000000
+1
+2000000
+1
+1000000
+2
+80800000
+2
+40400000
+2
+20200000
+2
+10100000
+2
+8080000
+2
+4040000
+2
+2020000
+2
+1010000
