+++ testResults/fullPawnTest.txt
@@ -1,0 +1,135 @@
+2
+808000000000
+
+00000000 
+00000000 
+10000000 
+10000000 
+00000000 
+00000000 
+00000000 
+00000000
+
+2
+404000000000
+2
+202000000000
+2
+101000000000
+2
+80800000000
+2
+40400000000
+2
+20200000000
+2
+10100000000
+1
+8000000000
+1
+4000000000
+1
+2000000000
+1
+1000000000
+1
+800000000
+1
+400000000
+1
+200000000
+1
+100000000
+1
+80000000
+1
+40000000
+1
+20000000
+1
+10000000
+1
+8000000
+1
+4000000
+1
+2000000
+1
+1000000
+1
+800000
+1
+400000
+1
+200000
+1
+100000
+1
+80000
+1
+40000
+1
+20000
+1
+10000
+1
+8000
+1
+4000
+1
+2000
+1
+1000
+1
+800
+1
+400
+1
+200
+1
+100
+1
+80
+1
+40
+1
+20
+1
+10
+1
+8
+1
+4
+1
+2
+1
+1
+
+
+
+00000001 
+00000000 
+00000000 
+00000000 
+00000000 
+00000000 
+00000000 
+00000000
+
+00000000 
+00000000 
+01000000 
+00000000 
+00000000 
+00000000 
+00000000 
+00000000
+
+00000000 
+00100000 
+00000000 
+00000000 
+00000000 
+00000000 
+00000000 
+00000000
