+++ testResults/pawnCaptureTest.txt
@@ -1,0 +1,51 @@
+3
+200000000000
+002000000000
+400000000000
+
+
+00000000 
+00100000 
+00000000 
+00000000 
+00000000 
+00000000 
+00000000 
+00000000
+
+00000000 
+00000000 
+01000000 
+00000000 
+00000000 
+00000000 
+00000000 
+00000000
+
+
+00000000 
+00000000 
+00100000 
+00000000 
+00000000 
+00000000 
+00000000 
+00000000
+
+00000000 
+00000000 
+00000000 
+00100000 
+00000000 
+00000000 
+00000000 
+00000000
+
+00000000 
+00000000 
+01000000 
+00000000 
+00000000 
+00000000 
+00000000 
+00000000
