+++ testResults/pawnCaptureTest2.txt
@@ -1,0 +1,28 @@
+3
+00000000
+00000000
+00100000
+00000000
+00000000
+00000000
+00000000
+00000000
+
+00000000
+00000000
+00000000
+00100000
+00000000
+00000000
+00000000
+00000000
+
+00000000
+00000000
+01000000
+00000000
+00000000
+00000000
+00000000
+00000000
+
